Fix bugs with last topic as player H5P, Image, Audio, Video or Project. Now user must click button to send progress and finish course. Refactor: Moved all bookmark notes to hook. Moved breakpoint functionality to hook. Moved progress requests of topic and h5p to hook. Moved CourseFinishModal to component. Moved course styles to component.
