+++ src/components/Course/CourseProgramContent.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useMemo, useCallback } from "react";
+import React, { useContext, useEffect, useMemo } from "react";
 import { EscolaLMSContext } from "@escolalms/sdk/lib/react/context";
 import { XAPIEvent } from "@escolalms/h5p-react";
 import TextPlayer from "./Players/TextPlayer";
@@ -29,58 +29,23 @@ export const CourseProgramContent: React.FC<{
   isThereAnotherTopic?: boolean;
   topicId?: number;
   preview?: boolean;
-  customNoCompletedEventsIds?: string[];
   disableNextTopicButton?: (b: boolean) => void;
+  onXAPI?: (event: XAPIEvent) => void;
 }> = ({
   lessonId,
   topicId,
   preview = false,
   disableNextTopicButton,
-  customNoCompletedEventsIds = [
-    "http://h5p.org/libraries/H5P.GuessTheAnswer-1.5",
-  ],
   isThereAnotherTopic = true,
+  onXAPI,
 }) => {
-  const { program, topicPing, topicIsFinished, sendProgress, h5pProgress } =
-    useContext(EscolaLMSContext);
+  const { program, topicPing, topicIsFinished } = useContext(EscolaLMSContext);
 
   const topic = useMemo(() => {
     return program.value?.lessons
       ?.find((lesson: API.Lesson) => lesson.id === lessonId)
       ?.topics?.find((topic: API.Topic) => topic.id === topicId);
   }, [program, lessonId, topicId]);
-
-  const onCompleteTopic = useCallback((): void => {
-    if (program?.value?.id) {
-      sendProgress(program?.value?.id, [
-        { topic_id: Number(topicId), status: 1 },
-      ]);
-    }
-  }, [program, topicId, sendProgress]);
-
-  const onXAPI = useCallback(
-    (event: XAPIEvent): void => {
-      isThereAnotherTopic &&
-        disableNextTopicButton &&
-        disableNextTopicButton(!Boolean(event?.statement?.verb?.id));
-
-      if (event?.statement) {
-        h5pProgress(
-          String(program?.value?.id),
-          Number(topicId),
-          event?.statement as API.IStatement
-        );
-      }
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [
-      program,
-      topicId,
-      h5pProgress,
-      disableNextTopicButton,
-      customNoCompletedEventsIds,
-    ]
-  );
 
   useEffect(() => {
     const isTopicFinished = topic?.id && topicIsFinished(topic.id);
@@ -113,12 +78,11 @@ export const CourseProgramContent: React.FC<{
     );
   }
   if (topic.topicable_type) {
-    // TODO: specific interface for advanced topic players -> example: ImagePlayer
     switch (topic.topicable_type) {
       case API.TopicType.H5P:
         return (
           <H5Player
-            onXAPI={(e: XAPIEvent) => onXAPI(e)}
+            onXAPI={(e: XAPIEvent) => onXAPI?.(e)}
             //@ts-ignore
             h5pObject={topic.topicable.content as API.H5PObject}
           />
@@ -133,18 +97,10 @@ export const CourseProgramContent: React.FC<{
         return <TextPlayer value={topic.topicable.value} />;
       case API.TopicType.Video:
         return (
-          <AudioVideoPlayer
-            mobile={isMobile}
-            url={topic.topicable.url}
-            light
-            onFinish={(): void => onCompleteTopic()}
-          />
+          <AudioVideoPlayer mobile={isMobile} url={topic.topicable.url} light />
         );
       case API.TopicType.Image:
         return <ImagePlayer topic={topic} onLoad={() => {}} />;
-      // Commented, we want complete topic on button click
-      // return <ImagePlayer topic={topic} onLoad={() => onCompleteTopic()} />;
-
       case API.TopicType.Audio:
         return (
           <AudioVideoPlayer
@@ -152,7 +108,6 @@ export const CourseProgramContent: React.FC<{
             audio
             url={topic.topicable.url}
             light
-            onFinish={(): void => onCompleteTopic()}
           />
         );
 
@@ -181,11 +136,7 @@ export const CourseProgramContent: React.FC<{
         return <GiftQuizPlayer topic={topic} />;
       case API.TopicType.Project:
         return (
-          <ProjectPlayer
-            course_id={program.value?.id ?? 0}
-            topic={topic}
-            onSuccess={onCompleteTopic}
-          />
+          <ProjectPlayer course_id={program.value?.id ?? 0} topic={topic} />
         );
       default:
         return <pre>{(topic as API.Topic).topicable_type}</pre>;
